FastAPI-Ecommerce-API-main/requirements.txt
alembic
annotated-types
anyio
bcrypt
click
colorama
dnspython
ecdsa
email-validator
exceptiongroup
fastapi
greenlet
h11
httptools
idna
Mako
MarkupSafe
passlib
psycopg2-binary
pyasn1
pydantic
pydantic-settings
pydantic_core
python-dotenv
python-jose
python-multipart
PyYAML
rsa
six
sniffio
SQLAlchemy
starlette
typing_extensions
uvicorn
watchfiles
websockets
httpx
supabase
